Move metatile_behaviors.h into include/constants
